==> JIM/01_1-Exploratory_Data_Analysis.ipynb <==
## EDA

import warnings
import numpy as np 
import pandas as pd
import seaborn as sns
import plotly.express as px
import matplotlib.pyplot as plt
from collections import Counter
from sklearn.feature_selection import f_classif
from sklearn.model_selection import train_test_split

warnings.filterwarnings('ignore')
df = pd.read_csv("data.csv", sep = ";")
pd.set_option('display.max_columns', None)
df.head()

df.groupby("GR").count()

df = df.loc[(df['GR'])!="D"]

df = df[['GR', 'RFV','H1RFV','H2RFV','CONICITY',
       'H1LFV', 'PLY','LFV','RRO', 'CAPSPLICE', 'H1RRO']]

X = df.drop([
'GR'
], axis=1)

X_all = X
y_all = df['GR']

X_train, X_test, y_train, y_test = train_test_split(X_all, y_all, test_size = 0.25 , stratify=y_all, random_state = 7265)

print('Numero de observaçoes do treino:', len(X_train))
print('Numero de observaçoes da teste:',len(X_test))

train = X_train.copy()
train['GR'] = y_train

test = X_test.copy()
test['GR'] = y_test

X_train.to_csv("X_train.csv", sep = ";", index=False)
X_test.to_csv("X_test.csv", sep = ";", index=False)
y_test.to_csv("y_test.csv", sep = ";", index=False)
y_train.to_csv("y_train.csv", sep = ";", index=False)

print(Counter(y_train).keys())
print(Counter(y_train).values())

print(Counter(y_test).keys())
print(Counter(y_test).values())

X_train.describe()

data = X_train.copy()
data['GR'] = y_train
data.head(2)

data.groupby("GR").describe()

plt.figure(figsize=(6,6))
data['GR'].value_counts().plot.pie(autopct="%1.1f%%")

g = sns.kdeplot(data["RFV"][(data["GR"] == "A") & (data["RFV"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["RFV"][(data["GR"] == "B") & (data["RFV"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["RFV"][(data["GR"] == "C") & (data["RFV"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("RFV")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["H1RFV"][(data["GR"] == "A") & (data["H1RFV"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["H1RFV"][(data["GR"] == "B") & (data["H1RFV"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["H1RFV"][(data["GR"] == "C") & (data["H1RFV"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("H1RFV")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["H2RFV"][(data["GR"] == "A") & (data["H2RFV"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["H2RFV"][(data["GR"] == "B") & (data["H2RFV"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["H2RFV"][(data["GR"] == "C") & (data["H2RFV"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("H2RFV")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["CONICITY"][(data["GR"] == "A") & (data["CONICITY"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["CONICITY"][(data["GR"] == "B") & (data["CONICITY"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["CONICITY"][(data["GR"] == "C") & (data["CONICITY"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("CONICITY")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["H1LFV"][(data["GR"] == "A") & (data["H1LFV"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["H1LFV"][(data["GR"] == "B") & (data["H1LFV"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["H1LFV"][(data["GR"] == "C") & (data["H1LFV"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("H1LFV")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["PLY"][(data["GR"] == "A") & (data["PLY"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["PLY"][(data["GR"] == "B") & (data["PLY"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["PLY"][(data["GR"] == "C") & (data["PLY"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("PLY")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["LFV"][(data["GR"] == "A") & (data["LFV"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["LFV"][(data["GR"] == "B") & (data["LFV"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["LFV"][(data["GR"] == "C") & (data["LFV"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("LFV")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["RRO"][(data["GR"] == "A") & (data["RRO"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["RRO"][(data["GR"] == "B") & (data["RRO"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["RRO"][(data["GR"] == "C") & (data["RRO"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("RRO")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["CAPSPLICE"][(data["GR"] == "A") & (data["CAPSPLICE"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["CAPSPLICE"][(data["GR"] == "B") & (data["CAPSPLICE"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["CAPSPLICE"][(data["GR"] == "C") & (data["CAPSPLICE"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("CAPSPLICE")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

g = sns.kdeplot(data["H1RRO"][(data["GR"] == "A") & (data["H1RRO"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["H1RRO"][(data["GR"] == "B") & (data["H1RRO"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["H1RRO"][(data["GR"] == "C") & (data["H1RRO"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("H1RRO")
g.set_ylabel("Density")
g = g.legend(["A","B", "C"])

warnings.filterwarnings('ignore')
orders = ["A", "B", "C"]
g = sns.factorplot(y="RFV",x="GR",data=data,kind="box", order = orders)
g = sns.factorplot(y="H1RFV",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="H2RFV",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="CONICITY",x="GR",data=data,kind="box", order = orders)
g = sns.factorplot(y="H1LFV",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="PLY",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="LFV",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="RRO",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="CAPSPLICE",x="GR", data=data,kind="box", order = orders)
g = sns.factorplot(y="H1RRO",x="GR", data=data,kind="box", order = orders)

Correlations

corrmat = np.triu(data.corr(method='pearson'))

data_cor = data[['RFV','H1RFV','H2RFV','CONICITY',
       'H1LFV', 'PLY','LFV','RRO', 'CAPSPLICE', 'H1RRO']]

g = sns.heatmap(data_cor.corr(method='pearson'),cmap="coolwarm",annot=True, mask = corrmat, fmt=".2f")

data_cor.corr(method='pearson')

top_num_features = data_cor.columns.to_list()
corrmat = data[top_num_features].corr()

plt.figure(figsize=(10,10))
sns.heatmap(corrmat);

### Mutual information

from sklearn.feature_selection import SelectKBest, SelectPercentile, mutual_info_classif

selector = SelectKBest(mutual_info_classif, k=7)
X_reduced = selector.fit_transform(X_train, y_train)
X_reduced.shape
cols = selector.get_support(indices=True)
selected_columns = X_train.iloc[:,cols].columns.tolist()
selected_columns

threshold = 10  # the number of most relevant features
high_score_features = []
feature_scores = mutual_info_classif(X_train, y_train, random_state=0)
for score, f_name in sorted(zip(feature_scores, X_train.columns), reverse=True)[:threshold]:
        print(f_name, score)
        high_score_features.append(f_name)
df_wine_norm_mic = X_train[high_score_features]
print(df_wine_norm_mic.columns)

ANOVA F-Statistic

F_statistic, p_values = f_classif(data_cor, y_train)
ANOVA_F_table = pd.DataFrame(data = {'Numerical_Feature': data_cor.columns.values, 'F-Score': F_statistic, 'p values': p_values.round(decimals=10)})
ANOVA_F_table.sort_values(by = ['F-Score'], ascending = False, ignore_index = True, inplace = True)
ANOVA_F_table

def Scatterplot(data, var1, var2, cat):

    fig = px.scatter(data, x = var1, y = var2, color =cat, width = 800)
    fig.update_traces(marker=dict(size = 12,line = dict(width = 1)),selector=dict(mode = 'markers'))
    fig.update_layout(title = 'Scatterplot: ' + var1 + " vs " + var2)
    fig.update_xaxes(title = var1)
    fig.update_yaxes(title = var2)
    fig.show()

Scatterplot(data, "RFV", "PLY",  "GR")

Scatterplot(data, "H1RFV", "CONICITY",  "GR")

Scatterplot(data, "H1RFV", "H2RFV",  "GR")

g = sns.pairplot(data, hue="GR", palette="Set2", diag_kind="kde", height=2.5)

fig = px.scatter_3d(data, x="H1RFV", y="CONICITY",z="RFV",
                    color= "GR")

fig.show()
