==> JIM/03_1-Models.ipynb <==
### Libraries

import time
import shap
import warnings
import numpy as np
import pandas as pd
import seaborn as sns
from collections import Counter
from IPython.display import Image
from sklearn.tree import export_graphviz
from yellowbrick.model_selection import FeatureImportances

from imblearn.over_sampling import SMOTE, ADASYN
from skopt.space import Real, Categorical, Integer

from sklearn.preprocessing import LabelEncoder, LabelBinarizer
from sklearn.preprocessing import StandardScaler, MinMaxScaler
import joblib
from plot_roc import *

from sklearn.metrics  import confusion_matrix, classification_report, roc_auc_score

warnings.filterwarnings('ignore')

from mlxtend.evaluate import cochrans_q
from mlxtend.evaluate import mcnemar_table
from mlxtend.evaluate import mcnemar

### Read 

warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

X_train = pd.read_csv("X_train.csv", sep = ";")
X_test = pd.read_csv("X_test.csv", sep = ";")

y_train = pd.read_csv("y_train.csv", sep = ";")
y_test = pd.read_csv("y_test.csv", sep = ";")

X_train.head()

## removendo variáveis 
X_train = X_train.drop([
'H1RFV', 'H1LFV', 'H1RRO'
], axis=1)

X_test = X_test.drop([
'H1RFV', 'H1LFV', 'H1RRO'
], axis=1)

def prepare_targets(y_train, y_test):
    le = LabelEncoder()
    le.fit(y_train)
    y_train_enc = le.transform(y_train)
    y_test_enc = le.transform(y_test)
    
    return y_train_enc, y_test_enc

y_train_enc, y_test_enc = prepare_targets(y_train, y_test)

y_train.GR.value_counts()

y_test.GR.value_counts()

y_test_enc

# y_test

X_test.head(5)

## Predict

def fit_and_print(model, X_test, y_enc):
    y_pred = model.predict(X_test)
    # print("Confusion Matrix: \n", confusion_matrix(y_test_enc, y_pred))
    sns.heatmap(confusion_matrix(y_enc, y_pred), annot=True)
    print("Classification Report: \n", classification_report(y_enc, y_pred))  


def fit_and_print_v2(model, X_test):
    y_pred = model.predict(X_test)
    cm = confusion_matrix(y_test_enc, y_pred)
    index =["A", "B", "C"]
    columns =["A", "B", "C"]
    cm_df = pd.DataFrame(cm,columns,index)                      
    plt.figure(figsize=(8,6))  
    sns.heatmap(cm_df, annot=True)
    print("Classification Report: \n", classification_report(y_test_enc, y_pred))  

    lb = LabelBinarizer()
    lb.fit(y_test_enc)
    y_test = lb.transform(y_test_enc)
    y_pred = lb.transform(y_pred)


    roc_auc = roc_auc_score(y_test, y_pred, multi_class = 'ovr', average=None)
    gini = 2*roc_auc -1
    print("Gini: ",gini)
    print("ROC AUC:: ",roc_auc)

file_minmax = "./models/minmax.pickle"
file_smote = "./models/smote.pickle"

minmax = joblib.load(file_minmax)
smote= joblib.load(file_smote)

X_train_minmax = minmax.transform(X_train)
X_test = minmax.transform(X_test)

X_resampled, y_resampled = smote.fit_resample(X_train_minmax, y_train_enc)

### Random Forest 

filename = './models/RF_best_bayes_resampled.sav'


RF_best = joblib.load(filename)
y_pred = RF_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

# fit_and_print(RF_best, X_test, y_test_enc)
fit_and_print_v2(RF_best, X_test)

prob_test = RF_best.predict_proba(X_test)[:,1]

data=pd.DataFrame()
data['GR'] = y_test.GR
data["Probability"] = RF_best.predict_proba(X_test)[:,0]
plt.xlim(.0, 1.)
g = sns.kdeplot(data["Probability"][(data["GR"] == "A") & (data["Probability"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "B") & (data["Probability"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "C") & (data["Probability"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("Probability")
g.set_ylabel("Count")
g = g.legend(["A","B", "C"])

plt.xlim(.0, 1.)

sns.kdeplot(
    data=data, x="Probability", hue="GR",
    cumulative=True, common_norm=False, common_grid=False, hue_order=['A', 'B', 'C'], palette=["Blue", "Green", 'Purple'],
)

# sns.histplot(data["prob"][(data["GR"] == "A") & (data["prob"].notnull())], kde = True, bins=30)
sns.histplot(data, x = "prob", kde = True, hue = "GR", bins=10)

data=pd.DataFrame()
data['GR'] = y_test.GR
data["Probability"] = RF_best.predict_proba(X_test)[:,1]
plt.xlim(.0, 1.)
g = sns.kdeplot(data["Probability"][(data["GR"] == "A") & (data["Probability"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "B") & (data["Probability"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "C") & (data["Probability"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("Probability")
g.set_ylabel("Count")
g = g.legend(["A","B", "C"])

plt.xlim(.0, 1.)

sns.kdeplot(
    data=data, x="Probability", hue="GR",
    cumulative=True, common_norm=False, common_grid=False, hue_order=['A', 'B', 'C'], palette=["Blue", "Green", 'Purple'],
)

data=pd.DataFrame()
data['GR'] = y_test.GR
data["Probability"] = RF_best.predict_proba(X_test)[:,2]
plt.xlim(.0, 1.)
g = sns.kdeplot(data["Probability"][(data["GR"] == "A") & (data["Probability"].notnull())], color="Blue", shade = True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "B") & (data["Probability"].notnull())], ax =g, color="Green", shade= True)
g = sns.kdeplot(data["Probability"][(data["GR"] == "C") & (data["Probability"].notnull())], ax =g, color="Purple", shade= True)
g.set_xlabel("Probability")
g.set_ylabel("Count")
g = g.legend(["A","B", "C"])

plt.xlim(.0, 1.)

sns.kdeplot(
    data=data, x="Probability", hue="GR",
    cumulative=True, common_norm=False, common_grid=False, hue_order=['A', 'B', 'C'], palette=["Blue", "Green", 'Purple'],
)

RF_best.estimators_[0]

tree_small_0 = RF_best.estimators_[0]
tree_small_1 = RF_best.estimators_[1]
feature_list = list(X_train.columns)
name_list = ['A', 'B', 'C']

export_graphviz(tree_small_0, out_file = 'tree_0_smote.dot', feature_names = feature_list, class_names = name_list,  rounded = True, precision = 1, filled = True)
export_graphviz(tree_small_1, out_file = 'tree_1_smote.dot', feature_names = feature_list, class_names = name_list,  rounded = True, precision = 1, filled = True)

# !dot -Tpng tree.dot -o small_tree.png -Gdpi=600

# Image(filename = 'images/small_tree_0_smote.png')

# Image(filename = 'images/small_tree_1_smote.png')

names_classifiers = [("Random Forest",RF_best)]
classifier = names_classifiers[0][1]
name = names_classifiers[0][0]
indices = np.argsort(classifier.feature_importances_)[::-1][:40]
g = sns.barplot(y=X_train.columns[indices][:40],x = classifier.feature_importances_[indices][:40] , orient='h')
g.set_xlabel("Relative importance",fontsize=12)
g.set_ylabel("Features",fontsize=12)
g.tick_params(labelsize=9)
g.set_title(name + " feature importance")

viz = FeatureImportances(RF_best, relative=False)
viz.fit(X_resampled, y_resampled)
viz.show()

explainer = shap.TreeExplainer(RF_best)
shap_values = explainer.shap_values(X_train_minmax)

# indices = np.argsort(shap_values)[::-1]
# X_train.columns[indices]

shap.summary_plot(shap_values, X_train_minmax, class_names= ['A', 'B', 'C'], feature_names = X_train.columns.tolist())

shap.summary_plot(shap_values[0], X_train_minmax, feature_names = X_train.columns.tolist())
shap.summary_plot(shap_values[1], X_train_minmax, feature_names = X_train.columns.tolist())
shap.summary_plot(shap_values[2], X_train_minmax, feature_names = X_train.columns.tolist())

for name in X_train.columns:
    shap.dependence_plot(name, shap_values[0], X_train, display_features=X_train)

for name in X_train.columns:
    shap.dependence_plot(name, shap_values[1], X_train, display_features=X_train)

for name in X_train.columns:
    shap.dependence_plot(name, shap_values[2], X_train, display_features=X_train)

shap.force_plot(explainer.expected_value[0], shap_values[0][1,:], X_train.iloc[1,:], matplotlib=True)
shap.force_plot(explainer.expected_value[1], shap_values[1][1,:], X_train.iloc[1,:], matplotlib=True)
shap.force_plot(explainer.expected_value[2], shap_values[2][1,:], X_train.iloc[1,:], matplotlib=True)

X_train.head(1)

y_train

# shap.initjs()
# shap.force_plot(explainer.expected_value[1], shap_values[1][:,:], X_resampled.iloc[:,:])

### Support Vector Machine

filename = './models/SVM_best_bayes_resampled.sav'

SVM_best = joblib.load(filename)
y_pred_svm = SVM_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred_svm)

fit_and_print(SVM_best, X_test, y_test_enc)

q, p_value = cochrans_q(y_test_enc, 
                       y_pred, 
                       y_pred_svm)

print('Q: %.3f' % q)
print('p-value: %.3f' % p_value)

from mlxtend.evaluate import combined_ftest_5x2cv

f, p = combined_ftest_5x2cv(estimator1=RF_best,
                            estimator2=SVM_best,
                            X=X_resampled, y=y_resampled,
                            random_seed=185585)

print('F statistic: %.3f' % f)
print('p value: %.3f' % p)

### Logistic Regression 

filename = './models/LR_best_bayes_resampled.sav'

LR_best = joblib.load(filename)
y_pred = LR_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(LR_best, X_test, y_test_enc)

fit_and_print(LR_best, X_train, y_train_enc)

### Multi-layer Perceptron

filename = './models/MLP_best_resampled.sav'

MLP_best = joblib.load(filename)
y_pred = MLP_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(MLP_best, X_test, y_test_enc)

fit_and_print(MLP_best, X_train, y_train_enc)

### KNN

filename = './models/KNN_best_bayes_resampled.sav'

KNN_best = joblib.load(filename)
y_pred = KNN_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(KNN_best, X_test, y_test_enc)

fit_and_print(KNN_best, X_train, y_train_enc)

### Gaussian Process

filename = './models/GP_best_bayes_resampled.sav'

GP_best = joblib.load(filename)
y_pred = GP_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(GP_best, X_test, y_test_enc)

fit_and_print(GP_best, X_train, y_train_enc)

### GBM

filename = './models/GBM_best_bayes_resampled.sav'

GBM_best = joblib.load(filename)
y_pred = GBM_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(GBM_best, X_test, y_test_enc)

### LGBM

filename = './models/LGBM_best_bayes_resampled.sav'

LGBM_best = joblib.load(filename)
y_pred = LGBM_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)


fit_and_print(LGBM_best, X_test, y_test_enc)

explainer = shap.TreeExplainer(LGBM_best)
shap_values = explainer.shap_values(X_resampled) 

shap.summary_plot(shap_values, X_resampled, class_names= ['A', 'B', 'C'])

shap.summary_plot(shap_values[0], X_resampled)
shap.summary_plot(shap_values[1], X_resampled)
shap.summary_plot(shap_values[2], X_resampled)

shap.force_plot(explainer.expected_value[0], shap_values[0][0,:], X_resampled.iloc[0,:], matplotlib=True)
shap.force_plot(explainer.expected_value[1], shap_values[1][0,:], X_resampled.iloc[0,:], matplotlib=True)
shap.force_plot(explainer.expected_value[2], shap_values[2][0,:], X_resampled.iloc[0,:], matplotlib=True)

shap.initjs()
shap.force_plot(explainer.expected_value[0], shap_values[0][:,:], X_resampled.iloc[:,:])

shap.initjs()
shap.force_plot(explainer.expected_value[1], shap_values[1][:,:], X_resampled.iloc[:,:])

shap.initjs()
shap.force_plot(explainer.expected_value[2], shap_values[2][:,:], X_resampled.iloc[:,:])

### XGBoost

filename = './models/XGB_best_bayes_resampled.sav'

XGB_best = joblib.load(filename)
y_pred = XGB_best.predict(X_test)
plot_roc_curve(y_test_enc, y_pred)

fit_and_print(XGB_best, X_test, y_test_enc)

explainer = shap.TreeExplainer(XGB_best)
shap_values = explainer.shap_values(X_resampled) 

shap.summary_plot(shap_values, X_resampled, class_names= ['A', 'B', 'C'])

shap.summary_plot(shap_values[0], X_resampled)
shap.summary_plot(shap_values[1], X_resampled)
shap.summary_plot(shap_values[2], X_resampled)
